feat: add rbac method in service

+++ services/auth-service/repository/rbac_repository.go
@@ -15,7 +15,7 @@ func NewRBACRepository(db *sql.DB) *RBACRepository {
 	}
 }
 
-func (r *RBACRepository) AssignRoleToUser(ctx context.Context, userId, roleId int64) error {
+func (r *RBACRepository) AssignRoleToUser(ctx context.Context, userId, roleId int) error {
 	query := `INSERT INTO user_roles (role_id, user_id) VALUES ($1, $2) RETURNING *;`
 	row := r.db.QueryRow(query, roleId, userId)
 	err := row.Scan(&roleId, &userId)
@@ -24,7 +24,7 @@ func (r *RBACRepository) AssignRoleToUser(ctx context.Context, userId, roleId in
 	}
 	return nil
 }
-func (r *RBACRepository) AssignPermissionToRole(ctx context.Context, permissionID, roleId int64) error {
+func (r *RBACRepository) AssignPermissionToRole(ctx context.Context, permissionID, roleId int) error {
 	query := `insert into role_permissions (role_id, permission_id) VALUES ($1, $2) RETURNING *;`
 	row := r.db.QueryRow(query, roleId, permissionID)
 	err := row.Scan(&roleId, &permissionID)
@@ -33,7 +33,7 @@ func (r *RBACRepository) AssignPermissionToRole(ctx context.Context, permissionI
 	}
 	return nil
 }
-func (r *RBACRepository) GetPermissionsByUserID(ctx context.Context, userID int64) ([]string, error) {
+func (r *RBACRepository) GetPermissionsByUserID(ctx context.Context, userID int) ([]string, error) {
 	query := `SELECT DISTINCT p.name From permissions p 
         join role_permissions rp on rp.permission_id = p.id
         join user_roles ur on ur.role_id = rp.role_id
